[Consensus] Move out pending votes to its own struct
Summary:
This is a purely technical noop refactoring, which is moving the pending votes from the larger structure of BlockTree to a separate PendingVotes struct.
We are going to update the logic of PendingVotes to generate TimeoutCertificates in a followup commits.

Testing:
Existing test coverage

Closes: #1132
Approved by: dmitri-perelman
