+++ consensus/src/chained_bft/block_storage/block_tree.rs
@@ -1,17 +1,17 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::chained_bft::block_storage::pending_votes::PendingVotes;
 use crate::{
     chained_bft::{
         block_storage::VoteReceptionResult,
-        common::Author,
         consensus_types::{block::ExecutedBlock, quorum_cert::QuorumCert, vote_msg::VoteMsg},
     },
     counters,
     util::time_service::duration_since_epoch,
 };
 use canonical_serialization::CanonicalSerialize;
-use crypto::{hash::CryptoHash, HashValue};
+use crypto::HashValue;
 use executor::StateComputeResult;
 use logger::prelude::*;
 use mirai_annotations::{checked_verify_eq, precondition};
@@ -22,10 +22,7 @@ use std::{
     sync::Arc,
     time::Duration,
 };
-use types::{
-    crypto_proxies::{LedgerInfoWithSignatures, ValidatorVerifier},
-    validator_verifier::VerifyError,
-};
+use types::crypto_proxies::ValidatorVerifier;
 
 /// This structure is a wrapper of [`ExecutedBlock`](crate::consensus_types::block::ExecutedBlock)
 /// that adds `children` field to know the parent-child relationship between blocks.
@@ -70,14 +67,6 @@ where
     }
 }
 
-/// This structure maintains tuple of block_id and LedgerInfo for last voted block by an Author
-/// We only remember latest vote from Author. Digest is used to identify and prune pending vote from
-/// same Author.
-struct BlockPendingVote {
-    block_id: HashValue,
-    digest: HashValue,
-}
-
 /// This structure maintains a consistent block tree of parent and children links. Blocks contain
 /// parent links and are immutable.  For all parent links, a child link exists. This structure
 /// should only be used internally in BlockStore.
@@ -93,17 +82,8 @@ pub struct BlockTree<T> {
     highest_quorum_cert: Arc<QuorumCert>,
     /// The quorum certificate that carries a highest ledger info
     highest_ledger_info: Arc<QuorumCert>,
-    /// `id_to_votes` might keep multiple LedgerInfos per proposed block in order
-    /// to tolerate non-determinism in execution: given a proposal, a QuorumCertificate is going
-    /// to be collected only for all the votes that carry identical LedgerInfo.
-    /// LedgerInfo digest covers the potential commit ids, as well as the vote information
-    /// (including the 3-chain of a voted proposal).
-    /// Thus, the structure of `id_to_votes` is as follows:
-    /// HashMap<proposed_block_id, HashMap<ledger_info_digest, LedgerInfoWithSignatures>>
-    id_to_votes: HashMap<HashValue, HashMap<HashValue, LedgerInfoWithSignatures>>,
-    /// Map of Author to last voted block id & digest. Any pending vote from Author is cleaned up
-    /// whenever new vote is added by same Author
-    author_to_last_voted_block_id: HashMap<Author, BlockPendingVote>,
+    /// Manages pending votes to be aggregated.
+    pending_votes: PendingVotes,
     /// Map of block id to its completed quorum certificate (2f + 1 votes)
     id_to_quorum_cert: HashMap<HashValue, Arc<QuorumCert>>,
     /// To keep the IDs of the elements that have been pruned from the tree but not cleaned up yet.
@@ -151,8 +131,7 @@ where
             highest_certified_block_id: root_id,
             highest_quorum_cert: Arc::clone(&root_quorum_cert),
             highest_ledger_info: Arc::new(root_ledger_info),
-            id_to_votes: HashMap::new(),
-            author_to_last_voted_block_id: HashMap::new(),
+            pending_votes: PendingVotes::new(),
             id_to_quorum_cert,
             pruned_block_ids,
             max_pruned_blocks_in_mem,
@@ -178,7 +157,6 @@ where
     fn remove_block(&mut self, block_id: HashValue) {
         // Remove the block from the store
         self.id_to_block.remove(&block_id);
-        self.id_to_votes.remove(&block_id);
         self.id_to_quorum_cert.remove(&block_id);
     }
 
@@ -302,99 +280,26 @@ where
         Ok(())
     }
 
-    /// Check if vote is valid. If this is the first vote from Author, add it to map. If Author has
-    /// already voted on same block then return DuplicateVote error. If Author has already voted
-    /// on some other block, prune last vote and insert new one in map.
-    fn check_vote_valid(&mut self, vote_msg: &VoteMsg) -> Result<(), VoteReceptionResult> {
-        let author = vote_msg.author();
-        let block_id = vote_msg.vote_data().block_id();
-        let digest = vote_msg.ledger_info().hash();
-
-        let last_voted_block = match self
-            .author_to_last_voted_block_id
-            .insert(author, BlockPendingVote { block_id, digest })
-        {
-            None => {
-                // First vote from Author, do nothing.
-                return Ok(());
-            }
-            Some(last_voted_block) => last_voted_block,
-        };
-
-        // Prune last pending vote from Author
-        if block_id == last_voted_block.block_id {
-            // Author has already voted for this block
-            return Err(VoteReceptionResult::DuplicateVote);
-        }
-
-        if let Some(block_pending_votes) = self.id_to_votes.get_mut(&last_voted_block.block_id) {
-            if let Some(li_digest_to_sig) = block_pending_votes.get_mut(&last_voted_block.digest) {
-                // Removing signature from last voted block
-                li_digest_to_sig.remove_signature(author);
-                if li_digest_to_sig.signatures().is_empty() {
-                    // Last vote/signature for block, remove digest entry
-                    block_pending_votes.remove(&last_voted_block.digest);
-                    if block_pending_votes.is_empty() {
-                        self.id_to_votes.remove(&last_voted_block.block_id);
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
     pub(super) fn insert_vote(
         &mut self,
         vote_msg: &VoteMsg,
         validator_verifier: Arc<ValidatorVerifier>,
     ) -> VoteReceptionResult {
-        let author = vote_msg.author();
         let block_id = vote_msg.vote_data().block_id();
         if let Some(old_qc) = self.id_to_quorum_cert.get(&block_id) {
             return VoteReceptionResult::OldQuorumCertificate(Arc::clone(old_qc));
         }
-
-        if let Err(e) = self.check_vote_valid(vote_msg) {
-            return e;
-        }
-
-        // All the votes collected for all the execution results of a given proposal.
-        let block_votes = self
-            .id_to_votes
-            .entry(block_id)
-            .or_insert_with(HashMap::new);
-
-        // Note that the digest covers the ledger info information, which is also indirectly
-        // covering vote data hash (in its `consensus_data_hash` field).
-        let digest = vote_msg.ledger_info().hash();
-        let li_with_sig = block_votes.entry(digest).or_insert_with(|| {
-            LedgerInfoWithSignatures::new(vote_msg.ledger_info().clone(), HashMap::new())
-        });
-
-        vote_msg.signature().clone().add_to_li(author, li_with_sig);
-        match validator_verifier.check_voting_power(li_with_sig.signatures().keys())
-        {
-            Ok(_) => {
-                let quorum_cert =
-                    QuorumCert::new(vote_msg.vote_data().clone(), li_with_sig.clone());
-                // Note that the block might not be present locally, in which case we cannot
-                // calculate time between block creation and qc
-                if let Some(time_to_qc) = self.get_block(&block_id).and_then(|block| {
-                    duration_since_epoch()
-                        .checked_sub(Duration::from_micros(block.timestamp_usecs()))
-                }) {
-                    counters::CREATION_TO_QC_S.observe_duration(time_to_qc);
-                }
-
-                VoteReceptionResult::NewQuorumCertificate(Arc::new(quorum_cert))
+        let res = self.pending_votes.insert_vote(vote_msg, validator_verifier);
+        if let VoteReceptionResult::NewQuorumCertificate(_) = res {
+            // Note that the block might not be present locally, in which case we cannot calculate
+            // time between block creation and qc
+            if let Some(time_to_qc) = self.get_block(&block_id).and_then(|block| {
+                duration_since_epoch().checked_sub(Duration::from_micros(block.timestamp_usecs()))
+            }) {
+                counters::CREATION_TO_QC_S.observe_duration(time_to_qc);
             }
-            Err(err) => match err {
-                VerifyError::TooLittleVotingPower { voting_power, .. } => {
-                    VoteReceptionResult::VoteAdded(voting_power)
-                }
-                _ => panic!("Impossible to get any other error except TooLittleVotingPower since messages with invalid authors are dropped, vote_msg = {}", vote_msg),
-            },
         }
+        res
     }
 
     /// Find the blocks to prune up to next_root_id (keep next_root_id's block). Any branches not
